Updated Modals to Support Forecast API
