+++ Weather/Apis/Forecast/ForecastApi.swift
@@ -1,22 +1,22 @@
 //
-//  DailyForecastApi.swift
+//  ForecastApi.swift
 //  Weather
 //
 //  Created by Kautsya Kanu on 22/02/21.
 //
 
 import Foundation
-class DailyForecastApi: BaseApiModal {
+class ForecastApi: BaseApiModal {
     //MARK:- Properties
     private var cityName: String = ""
-    private var numberOfDays: Int = 7
+    private var numberOfTimeStamps: Int = 8
     private var units: MeasurementUnit = .standard
     private var language: Language = .english
     var cityForecast: ForecastWeather?
     //MARK:- Fetchable
-    override var apiEndPoint: String { "data/2.5/daily" }
+    override var apiEndPoint: String { "data/2.5/forecast" }
     override var params: [String : AnyHashable] {
-        ["q": cityName, "cnt": numberOfDays, "units": units, "lang": language.code]
+        ["q": cityName, "cnt": numberOfTimeStamps, "units": units, "lang": language.code]
     }
     override func parse(_ data: Data, for params: [String : AnyHashable]) throws {
         let decoder = JSONDecoder()
