+++ Weather/Apis/Forecast/ForecastWeather.swift
@@ -15,29 +15,19 @@ struct ForecastWeather: Codable {
     let list: [WeatherDayDetail]
 }
 
+//MARK:- Temporary Structures
 extension ForecastWeather {
     struct WeatherDayDetail: Codable {
-        ///Time of data forecasted
+        ///Time of data forecasted, unix, UTC
         let dt: Int
-        let sunrise: Int
-        let sunset: Int
-        let temp: Temperature
-        let feelsLike: Temperature
-        ///Atmospheric pressure on the sea level, **hPa**
-        let pressure: Double
-        ///Humidity **%**
-        let humidity: Int
+        let main: Main
         let weather: [Weather]
-        ///Wind speed
-        /// - **Unit Default**: meter/sec
-        /// - **Metric**: meter/sec
-        /// - **Imperial**: miles/hour.
-        let speed: Float
-        ///Wind direction, degrees (meteorological)
-        let deg: Double
         ///Cloudiness **%**
-        let clouds: Int
+        let clouds: Cloud
+        let wind: Wind
+        let visibility: Int
         ///Probability of precipitation
         let pop: Float
+        let rain: Rain
     }
 }
