+++ Weather/Modals/Cloud.swift
@@ -1,0 +1,11 @@
+//
+//  Cloud.swift
+//  Weather
+//
+//  Created by Kautsya Kanu on 23/02/21.
+//
+
+import Foundation
+struct Cloud: Codable {
+    let all: Int
+}
